# server/common/swagger/Api.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: express-microservice
  description: Express Microservice Starter
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /examples/{id}:
    get:
      tags:
        - Examples
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the entity to retrieve
          type: integer
      responses:
        200:
          description: Return the example with the specified id
        404:
          description: Example not available
...